move changes to cube.py for JCMT to new folder
